+++ CHANGELOG.rst
@@ -5,6 +5,4 @@ Changelog
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Added Exact Riemann solver for the 1-D Euler equations